Initial refactor for Accumulo 2.0
